feat: AttackArea node, everything the UnitAttack script should need made
